requirements.txt
absl-py==0.2.2
astor==0.6.2
audioread==2.1.6
backcall==0.1.0
bleach==1.5.0
certifi==2018.4.16
colorama==0.3.9
cycler==0.10.0
Cython==0.28.3
decorator==4.3.0
entrypoints==0.2.3
gast==0.2.0
grpcio==1.12.1
html5lib==0.9999999
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.0
Jinja2==2.10
joblib==0.12.0
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
kiwisolver==1.0.1
librosa==0.6.1
llvmlite==0.23.2
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
mkl-fft==1.0.0
mkl-random==1.0.1
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.5.0
numba==0.38.1
numpy==1.14.5
pafy==0.5.4
pandocfilters==1.4.2
parso==0.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
protobuf==3.6.0
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.3
python-mnist==0.5
pytz==2018.4
pywinpty==0.5.4
pyzmq==17.0.0
resampy==0.2.1
scikit-learn==0.19.1
scipy==1.1.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
tensorboard==1.8.0
tensorflow-gpu==1.8.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tornado==5.0.2
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.14.1
win-unicode-console==0.5
wincertstore==0.2
youtube-dl==2018.6.25
